+++ mds/apis/prv_api/providers.py
@@ -6,7 +6,13 @@ from mds.access_control.permissions import require_scopes
 from mds.access_control.scopes import SCOPE_PRV_API
 
 
+class ProviderColorsSerializer(serializers.Serializer):
+    primary = serializers.CharField(required=False)
+    secondary = serializers.CharField(required=False)
+
+
 class ProviderSerializer(serializers.ModelSerializer):
+
     id = serializers.UUIDField(
         required=True, help_text="Unique provider identifier (UUID)"
     )
@@ -14,12 +20,21 @@ class ProviderSerializer(serializers.ModelSerializer):
     logo_b64 = serializers.CharField(
         required=False, help_text="Logo of provider base64 encoded"
     )
+    colors = ProviderColorsSerializer(
+        default=dict, help_text="colors for distinguishing providers on map"
+    )
+
+    def create(self, validated_data):
+        instance = self.Meta.model(**validated_data)
+        instance.save()
+        return instance
 
     class Meta:
         model = models.Provider
         fields = (
             "id",
             "name",
+            "colors",
             "logo_b64",
             "base_api_url",
             "device_category",